prepare for salsa20 sse2
