+++ salsa20/src/lib.rs
@@ -76,16 +76,17 @@
 pub use cipher;
 
 use cipher::{
-    consts::{U1, U10, U24, U32, U4, U6, U64, U8},
+    consts::{U10, U24, U32, U4, U6, U64, U8},
     generic_array::{typenum::Unsigned, GenericArray},
-    Block, BlockSizeUser, IvSizeUser, KeyIvInit, KeySizeUser, ParBlocksSizeUser, StreamBackend,
-    StreamCipherCore, StreamCipherCoreWrapper, StreamCipherSeekCore, StreamClosure,
+    Block, BlockSizeUser, IvSizeUser, KeyIvInit, KeySizeUser, StreamCipherCore,
+    StreamCipherCoreWrapper, StreamCipherSeekCore, StreamClosure,
 };
 use core::marker::PhantomData;
 
 #[cfg(feature = "zeroize")]
 use cipher::zeroize::{Zeroize, ZeroizeOnDrop};
 
+mod backends;
 mod xsalsa;
 
 pub use xsalsa::{hsalsa, XSalsa12, XSalsa20, XSalsa8, XSalsaCore};
@@ -189,7 +190,7 @@ impl<R: Unsigned> StreamCipherCore for SalsaCore<R> {
         rem.try_into().ok()
     }
     fn process_with_backend(&mut self, f: impl StreamClosure<BlockSize = Self::BlockSize>) {
-        f.call(&mut Backend(self));
+        f.call(&mut backends::soft::Backend(self));
     }
 }
 
@@ -219,64 +220,3 @@ impl<R: Unsigned> Drop for SalsaCore<R> {
 #[cfg(feature = "zeroize")]
 #[cfg_attr(docsrs, doc(cfg(feature = "zeroize")))]
 impl<R: Unsigned> ZeroizeOnDrop for SalsaCore<R> {}
-
-struct Backend<'a, R: Unsigned>(&'a mut SalsaCore<R>);
-
-impl<'a, R: Unsigned> BlockSizeUser for Backend<'a, R> {
-    type BlockSize = U64;
-}
-
-impl<'a, R: Unsigned> ParBlocksSizeUser for Backend<'a, R> {
-    type ParBlocksSize = U1;
-}
-
-impl<'a, R: Unsigned> StreamBackend for Backend<'a, R> {
-    #[inline(always)]
-    fn gen_ks_block(&mut self, block: &mut Block<Self>) {
-        let res = run_rounds::<R>(&self.0.state);
-        self.0.set_block_pos(self.0.get_block_pos() + 1);
-
-        for (chunk, val) in block.chunks_exact_mut(4).zip(res.iter()) {
-            chunk.copy_from_slice(&val.to_le_bytes());
-        }
-    }
-}
-
-#[inline]
-#[allow(clippy::many_single_char_names)]
-pub(crate) fn quarter_round(
-    a: usize,
-    b: usize,
-    c: usize,
-    d: usize,
-    state: &mut [u32; STATE_WORDS],
-) {
-    state[b] ^= state[a].wrapping_add(state[d]).rotate_left(7);
-    state[c] ^= state[b].wrapping_add(state[a]).rotate_left(9);
-    state[d] ^= state[c].wrapping_add(state[b]).rotate_left(13);
-    state[a] ^= state[d].wrapping_add(state[c]).rotate_left(18);
-}
-
-#[inline(always)]
-fn run_rounds<R: Unsigned>(state: &[u32; STATE_WORDS]) -> [u32; STATE_WORDS] {
-    let mut res = *state;
-
-    for _ in 0..R::USIZE {
-        // column rounds
-        quarter_round(0, 4, 8, 12, &mut res);
-        quarter_round(5, 9, 13, 1, &mut res);
-        quarter_round(10, 14, 2, 6, &mut res);
-        quarter_round(15, 3, 7, 11, &mut res);
-
-        // diagonal rounds
-        quarter_round(0, 1, 2, 3, &mut res);
-        quarter_round(5, 6, 7, 4, &mut res);
-        quarter_round(10, 11, 8, 9, &mut res);
-        quarter_round(15, 12, 13, 14, &mut res);
-    }
-
-    for (s1, s0) in res.iter_mut().zip(state.iter()) {
-        *s1 = s1.wrapping_add(*s0);
-    }
-    res
-}
